+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,10 +1,9 @@
 package roomescape.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationCreateRequest;
 import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationTimeCreateRequest;
-import roomescape.dto.ReservationTimeResponse;
 import roomescape.entity.ReservationTime;
 import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
@@ -16,6 +15,7 @@ public class ReservationService {
     private final ReservationDao reservationDao;
     private final ReservationTimeDao reservationTimeDao;
 
+    @Autowired
     public ReservationService(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao) {
         this.reservationDao = reservationDao;
         this.reservationTimeDao = reservationTimeDao;
@@ -33,18 +33,5 @@ public class ReservationService {
 
     public void deleteReservation(Long id) {
         reservationDao.deleteById(id);
-    }
-
-    public ReservationTimeResponse createReservationTime(ReservationTimeCreateRequest request) {
-        Long id = reservationTimeDao.save(request.toEntity());
-        return ReservationTimeResponse.fromEntity(reservationTimeDao.findById(id));
-    }
-
-    public List<ReservationTimeResponse> readAllReservationTime() {
-        return ReservationTimeResponse.fromEntities(reservationTimeDao.findAll());
-    }
-
-    public void deleteReservationTime(Long id) {
-        reservationTimeDao.deleteById(id);
     }
 }
