+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,0 +1,32 @@
+package roomescape.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import roomescape.dto.ReservationTimeCreateRequest;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.repository.ReservationTimeDao;
+
+import java.util.List;
+
+@Service
+public class ReservationTimeService {
+    private final ReservationTimeDao reservationTimeDao;
+
+    @Autowired
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
+    }
+
+    public ReservationTimeResponse createReservationTime(ReservationTimeCreateRequest request) {
+        Long id = reservationTimeDao.save(request.toEntity());
+        return ReservationTimeResponse.fromEntity(reservationTimeDao.findById(id));
+    }
+
+    public List<ReservationTimeResponse> readAllReservationTime() {
+        return ReservationTimeResponse.fromEntities(reservationTimeDao.findAll());
+    }
+
+    public void deleteReservationTime(Long id) {
+        reservationTimeDao.deleteById(id);
+    }
+}
